- Added unsafe memcpy and strlen functions checking whether mappings are correct - Added helper to validate addresses for syscalls - Replaced usage of memcpy by memcpy_unsafe within system calls - Added usage of strlen_unsafe with validation where strings are used - Revised kernel puts system call to duplicate string to print instead of direct usage - Revised elf_image_size to consider address as unsafe - Revised duplicate function to consider src as unsafe - Added some additional validation checks to process replace system call

+++ driver/src/generic/core/startup/main.c
@@ -230,7 +230,7 @@ ssize_t _write(
   size_t len
 ) {
   if ( file == STDOUT_FILENO || file == STDERR_FILENO ) {
-    return _kernel_puts( ( char* )ptr, ( int )len );
+    return _kernel_puts( ( char* )ptr, len );
   }
   return -1;
 }
